Corrected ColorPreference javadoc and moved to ui.color
