+++ app/src/main/java/com/amaze/filemanager/ui/colors/ColorPreferenceHelper.java
@@ -8,7 +8,7 @@ import android.support.annotation.ColorRes;
 import com.amaze.filemanager.R;
 import com.amaze.filemanager.fragments.preference_fragments.PreferencesConstants;
 import com.amaze.filemanager.utils.Utils;
-import com.amaze.filemanager.utils.color.ColorPreference;
+import com.amaze.filemanager.ui.colors.ColorPreference;
 
 import java.util.Arrays;
 import java.util.List;
